refactor running script into main.py
